authable v0.3.2 Authable.Token

Oauth2 token store

Functions

access_token_changeset(model, params \\ :empty)

authorization_code_changeset(model, params \\ :empty)

changeset(model, params \\ :empty)

Creates a changeset based on the model and params.

If no params are provided, an invalid changeset is returned with no validation performed.

is_expired?(token)

refresh_token_changeset(model, params \\ :empty)

session_token_changeset(model, params \\ :empty)